+++ kafka-client/src/main/java/dev/responsive/kafka/clients/EndOffsetsPoller.java
