+++ kafka-client/src/test/java/dev/responsive/kafka/clients/EndOffsetsPollerTest.java
