+++ kafka-client/src/test/java/dev/responsive/kafka/clients/ResponsiveGlobalConsumerTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.responsive.kafka.consumer;
+package dev.responsive.kafka.clients;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
@@ -25,6 +25,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
+import dev.responsive.kafka.clients.ResponsiveGlobalConsumer;
 import java.time.Duration;
 import java.util.List;
 import java.util.Map;
